requirements.txt
python-telegram-bot==20.7
flask==3.0.0
google-genai